requirements.txt
pyyaml
discord
ffmpeg
pyttsx3
PySimpleGUI